requirements.txt
aiofiles==24.1.0
ImageHash==4.3.1
numpy==2.0.1
pillow==10.4.0
PyWavelets==1.6.0
scipy==1.14.0
